v4l/Makefile: accept alternate EXTRAVERSION
Some OOT kernels use a numeric value for EXTRAVERSION like:
	VERSION = 3
	ATCHLEVEL = 4
	SUBLEVEL = 39
	EXTRAVERSION = 13
	NAME = Saber-toothed Squirrel

The current logic would make the Kernel version to be:

	VERSION=3
	PATCHLEVEL:=4
	SUBLEVEL:=39
	KERNELRELEASE:=3.4.3913
	SRCDIR:=/tmp/sandbox/custom-linux-3.4

In the above example, as 3913 is bigger than 256, the KERNEL_VERSION
macro would compute a wrong value for the version number, causing it
to be misinterpreted as Kernel 3.19.73.

Fix it by making the logic to represent the above as:
	KERNELRELEASE:=3.4.39.13
if the EXTRAVERSION is a pure numeric value.

Reported-by: Mason <slash.tmp@free.fr>
Signed-off-by: Mauro Carvalho Chehab <mchehab@osg.samsung.com>