only one select Kconfig needs to match
When a driver has multiple lines like this
(with is the case of DVB_USB_CXUSB Kconfig logic):
        select DVB_SI2168 if MEDIA_SUBDRV_AUTOSELECT

The make_kconfig.pl script transforms those into dependencies, but
if a symbol like that is not backward compatible with the current
build version, it will unselect the entire driver.

Relax the logic to still compile the driver, if one or more such
autoselects happen.

Tested-by:  Vincent McIntyre <vincent.mcintyre@gmail.com>
Signed-off-by: Mauro Carvalho Chehab <mchehab@s-opensource.com>