add fake code for usb_urb_ep_type_check()
The usb_urb_ep_type_check() function checks if an endpoint
is valid, for security reasons, in order to avoid bad hardware
to be probed.

While it is interesting to support it on backward versions,
media drivers work just fine without that. Also, on older
Kernels, such test doesn't exist.

So, it should not be less safe to just return zero on
this function, on legacy Kernels.

Reported-by: Vincent McIntyre <vincent.mcintyre@gmail.com>
Signed-off-by: Mauro Carvalho Chehab <mchehab@s-opensource.com>