Fix frame vector wildcard file check
This check was consistently failing on all systems tested.
The path to object directory is used here to explicitly override
CWD. The thought is, if frame_vector.c exists in the build
directory then the build system has determined it is required,
and the source therefore should be compiled. The module will
not be built unless the build system has enabled it's config
option anyways, so this change should be safe in all circumstances.

Signed-off-by: Brad Love <brad@nextdimension.cc>
Signed-off-by: Hans Verkuil <hansverk@cisco.com>