Fix spurious config-mycompat.h warnings
Apparently a zero byte file size leads to "could not open" warnings.
Adding a single comment to the file removes the warnings displayed
after compilation.

To prevent overwriting a user supplied config-mycompat.h the file
is only created if it does not already exist.

Also fix for possible spaces in the path.

Works without warning when using both the do-it-all build script
as well as manual operation via make targets.

Signed-off-by: Brad Love <brad@nextdimension.cc>
Signed-off-by: Jasmin Jessich <jasmin@anw.at>