Disabled VIDEO_CROS_EC_CEC for Kernels older than 4.19
The CEC driver for ChromeOS Embedded Controller requires union
ec_response_get_next_data_v1 with member cec_message, which is
available only since Kernel 4.19.
It is a very new driver, so it is ok to disable it for 4.18 and
below.
If someone needs this driver for 4.18 an below, please provide
a patch to get it compiled.

Signed-off-by: Jasmin Jessich <jasmin@anw.at>