make_makefile.pl: ensure that compressed old modules will be removed
The install logic doesn't do the right thing with compressed files.
They'll be kept until make rminstall is called. Also, it will only
remove files that are gzipped. However, nowadays, distros are packaging
files also with *.xz and other formats.

So, make the module removal more complete and call it at make install
too.

That makes it a way more verbose, so change the logic a little bit in
order to not display anything when directories aren't found.

Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>