media-build: build against a specific kernel
I find it helpful to be able to build the linux-media modules
for a kernel other than the one I am currently running.

This is a resend of the patch that in two previous attempts
made it into patchwork but the text was line-wrapped.
Those patches should be dropped.
  https://patchwork.linuxtv.org/patch/59979/
  https://patchwork.linuxtv.org/patch/60186/

There are no differences in the patch, just the submission email.
Apologies for all the noise.

Signed-off-by: Vincent McIntyre <media@arewethere.net>
Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>